% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_manipulation.R
\name{add_identical_lists}
\alias{add_identical_lists}
\title{Add identical lists
This function adds together two identical lists. In the case where either is empty it adds 0. If the lists are of unequal length, it will return the greatest list possible such that the two entries are added together.}
\usage{
add_identical_lists(x, y)
}
\arguments{
\item{y}{}
}
\description{
Add identical lists
This function adds together two identical lists. In the case where either is empty it adds 0. If the lists are of unequal length, it will return the greatest list possible such that the two entries are added together.
}
\examples{
add_identical_lists(x = list_structure[[index_name]][[second_level_indexes[[index_position]]]],y = return_value)
}
