% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability_functions.R
\name{beta_mom}
\alias{beta_mom}
\title{beta_mom}
\usage{
beta_mom(data)
}
\arguments{
\item{data}{Data frame.}
}
\value{
A vector of estimated parameters for the Beta distribution.
}
\description{
This function estimates the parameters of the Beta distribution which best fits the data, using the method of moments approach.
}
\examples{
mom_fit_beta <- beta_mom(sub_dat$value)
}
