% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_manipulation.R
\name{calc_scale_factors}
\alias{calc_scale_factors}
\title{Calculate Scaling Factors}
\usage{
calc_scale_factors(scaling_data, chosen_scale)
}
\arguments{
\item{chosen_scale}{}
}
\value{
Data frame
}
\description{
This function takes a set of scaling data for a chosen country and turns it into factors relative to the most recent year
}
\examples{
scaling_data <- calc_scale_factors(raw_scaling_data, chosen_scale)
}
