% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_track.R
\name{calculate_convolve_spines}
\alias{calculate_convolve_spines}
\title{Title}
\usage{
calculate_convolve_spines(track, frame_rate = 16)
}
\arguments{
\item{track}{}

\item{frame_rate}{the frame rate of the tracked videos}
}
\value{

}
\description{
Title
}
