% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_help.R
\name{clockwise_angle_v}
\alias{clockwise_angle_v}
\title{Clockwise angle vectorized}
\usage{
clockwise_angle_v(a1, a2, b1, b2)
}
\arguments{
\item{a1}{vector}

\item{a2}{vector}

\item{b1}{vector}

\item{b2}{vector}
}
\value{
clockwise angle between vector a and b
}
\description{
Calculates clockwise angle between two vectors.
Positive means clockwise change and negative counter clockwise.
}
