% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear_detrending.R
\name{detrend_linear_data}
\alias{detrend_linear_data}
\title{detrend_linear_data}
\usage{
detrend_linear_data(dat)
}
\arguments{
\item{dat}{Data frame with linear trend.}
}
\value{
Data frame with detrended data.
}
\description{
This function applies linear detrending to the data.
}
\examples{
sub_data <- detrend_linear_data(sub_data)
}
