% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability_functions.R
\name{dfrechet}
\alias{dfrechet}
\title{dfrechet}
\usage{
dfrechet(x, alpha, s)
}
\arguments{
\item{x}{Vector.}

\item{alpha}{Shape parameter.}

\item{s}{Scale parameter.}
}
\value{
Density function.
}
\description{
Density function for the Frechet distribution with shape and scale parameters.
}
\examples{
dfrechet(2:4, 0.8, 0.5)
}
