% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_manipulation.R
\name{elongate_core_data}
\alias{elongate_core_data}
\title{Elongate core data}
\usage{
elongate_core_data(cdx)
}
\arguments{
\item{cdx}{}
}
\value{
Data frame
}
\description{
Transform core data back to long format
}
\examples{
cdx1 <- cdx[[1]]
cdx2 <- cdx[[2]]
cdx1 <- widen_core_data(cdx1)
cdx1[i,j+1] <- as.numeric(v)
cdx1 <- elongate_core_data(cdx1)
cdx2 <- values_to_frequency(cdx1)
}
