% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_manipulation.R
\name{expand_data}
\alias{expand_data}
\title{Expand data}
\usage{
expand_data(data, country_archetype)
}
\arguments{
\item{data}{}
}
\value{
Data frame.
}
\description{
This function takes an arbitrary data set and expands from long form to wide form.
}
\examples{
expand_data(cost_data, "country")
expand_data(damage_data, "country")
}
