% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution_fitting.R
\name{get_aic_mle}
\alias{get_aic_mle}
\title{Get AIC and MLE parameters}
\usage{
get_aic_mle(dat, updateProgress = NULL, data_type = "Historical", years)
}
\arguments{
\item{updateProgress}{}
}
\value{
Data frame with all the distribution fits for each peril.
}
\description{
This function takes the data and fits a distribution for each peril and returns the Maximum likelihood estimates of the parameters, along with the AIC score.
This also returns (in the case of a confidence interval fit) 1000 MLE fits. Used within the fit_distribution function.
}
\examples{
out <- get_aic_mle(the_right_data, updateProgress)
}
