% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{make_simulation}
\alias{make_simulation}
\title{make_simulation}
\usage{
make_simulation(dis_name, dat, n_sim)
}
\arguments{
\item{dis_name}{The name of the distribution, which will used to run the simulations.}

\item{dat}{Data frame which includes the corresponding MLE values.}
}
\value{
15,000 simulated values.
}
\description{
This function takes the name of the distribution with the lowest AIC score and the corresponding MLE and AIC values, then simulates 15,000 observations.
}
\examples{
make_simulation(dis_name = sub_peril$distribution, dat = sub_peril)
}
