% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_manipulation.R
\name{match_dist_names}
\alias{match_dist_names}
\title{Match distribution names}
\usage{
match_dist_names(x)
}
\arguments{
\item{x}{}
}
\description{
Matches distribution names such that if they are NULL or NA, then NULL or NA is returned. If the distribution names are one of: 'lnorm', 'Beta_ab', 'gamma','frechet', 'gumbel', 'weibull','pareto', then consequently 'Log normal', 'Beta', 'Gamma', 'Frechet', 'Gumbel', 'Weibull', 'Pareto' is returned.
}
\examples{
peril_dists <- lapply(list(input$dist_drought_input,input$dist_earthquake_input,input$dist_flood_input,input$dist_storm_input),FUN=match_dist_names)
}
