% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability_functions.R
\name{pBeta_ab}
\alias{pBeta_ab}
\title{pBeta_ab}
\usage{
pBeta_ab(q, shape1 = 2, shape2 = 3, b = 1)
}
\arguments{
\item{q}{A vector.}

\item{shape1}{Shape parameter.}

\item{shape2}{Shape parameter.}

\item{b}{Boundary parameter.}
}
\value{
Distribution function.
}
\description{
Distribution function for the 4-parameter Beta distribution.
}
