% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability_functions.R
\name{pfrechet}
\alias{pfrechet}
\title{pfrechet}
\usage{
pfrechet(q, alpha, s)
}
\arguments{
\item{q}{Vector.}

\item{alpha}{Shape parameter.}

\item{s}{Scale parameter.}
}
\value{
Distribution function.
}
\description{
Distribution function for the Frechet distribution with shape and scale parameters.
}
\examples{
pfrechet(2:4, 0.8, 0.5)
}
