% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability_functions.R
\name{pgumbel}
\alias{pgumbel}
\title{pgumbel}
\usage{
pgumbel(q, mu, s)
}
\arguments{
\item{q}{Vector.}

\item{mu}{Location parameter.}

\item{s}{Scale parameter.}
}
\value{
Distribution function.
}
\description{
Distribution function for the Gumbel distribution with location and scale parameters.
}
\examples{
pgumbel(-1:2, -1, 0.5)
}
