% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability_functions.R
\name{qBeta_ab}
\alias{qBeta_ab}
\title{qBeta_ab}
\usage{
qBeta_ab(p, shape1 = 2, shape2 = 3, b = 1)
}
\arguments{
\item{p}{A vector of probabilities.}

\item{shape1}{Shape parameter.}

\item{shape2}{Shape parameter.}

\item{b}{Boundary parameter.}
}
\value{
Quantile function.
}
\description{
Quantile function for the 4-parameter Beta distribution.
}
