% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output_functions.R
\name{quant_that}
\alias{quant_that}
\title{quant_that}
\usage{
quant_that(dat_sim, dat = NULL, combined_ci = NULL)
}
\arguments{
\item{dat_sim}{Data frame of simulated values from the run_simulations function.}

\item{dat}{Dataframe combines the original loss data based on the perils selected.}

\item{combined_ci}{}
}
\value{
Data frame for Output 1.
}
\description{
Function that creates the Output data frame for Estimates of annual loss with Long-term average, 1 in 5 years, 1 in 10 years, 1 in 25 years, 1 in 50 years, 1 in 100 years, Highest historical annual loss and Most recent annual loss.
}
\examples{
sub_dat <- quant_that(dat_sim = dat_sim,
dat = dat,
combined_ci = perils_ci)
sub_dat$variable <- factor(sub_dat$variable, levels = c('1 in 5 Years', '1 in 10 Years', '1 in 25 Years', '1 in 50 Years', '1 in 100 Years',
                                                        'Long-term average', 'Highest historical annual loss', 'Most recent annual loss'))
}
