% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability_functions.R
\name{rbern}
\alias{rbern}
\title{rbern}
\usage{
rbern(n, prob)
}
\arguments{
\item{n}{Sample size.}

\item{prob}{A vector of success probabilities.}
}
\value{
Binary vector of length \code{n}.
}
\description{
Random sample from Bernoulli distribution.
}
\examples{
rbern(1000, 0.5)
rbern(5000, 0.04)
}
