% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability_functions.R
\name{rBeta_ab}
\alias{rBeta_ab}
\title{rBeta_ab}
\usage{
rBeta_ab(n, shape1 = 2, shape2 = 3, b = 1)
}
\arguments{
\item{n}{Number of observations.}

\item{shape1}{Shape parameter.}

\item{shape2}{Shape parameter.}

\item{b}{Boundary parameter.}
}
\value{
Binary vector of length \code{n}.
}
\description{
Generates random deviates from the 4-parameter Beta distribution.
}
