% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability_functions.R
\name{rgumbel}
\alias{rgumbel}
\title{rgumbel
Generate random deviates from the Gumbel distribution with location and scale parameters.}
\usage{
rgumbel(n, mu, s)
}
\arguments{
\item{n}{Number of observations.}

\item{mu}{Location parameter.}

\item{s}{Scale parameter.}
}
\value{
Binary vector of length \code{n}.
}
\description{
rgumbel
Generate random deviates from the Gumbel distribution with location and scale parameters.
}
\examples{
rgumbel(6, -1, 0.5)
}
