% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_help.R
\name{rotate_vector_clockwise}
\alias{rotate_vector_clockwise}
\title{Rotate vector clockwise}
\usage{
rotate_vector_clockwise(vector, angle)
}
\arguments{
\item{vector}{2d vector}

\item{angle}{angle in radiants}
}
\value{
rotated 2d vector
}
\description{
rotates 2d vector in clockwise direction
}
