% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{run_simulations}
\alias{run_simulations}
\title{run_simulations}
\usage{
run_simulations(
  prepared_simulation_data = NULL,
  prepared_frequency_data = NULL
)
}
\arguments{
\item{prepared_simulation_data}{Data frame that comes from the prepare_simulations function.}

\item{prepared_frequency_data}{Data frame of the simulated frequency data that comes from sim_bern function.}
}
\value{
A data frame with 15,000 rows with simulated values, including simulated values for both upper and lower quartiles, the simulated frequencies, and then the outcomes, including the upper and lower quartiles.
}
\description{
This function checks if the data is ready for simulation, then simulates the data with the "make_simulations" function. The simulated loss data is then multiplied by the simulated frequency data.
}
\examples{
ran_simulations <- reactive({
rep_bern <- repeatable(simulate_bernoulli, seed = 122)
bs <- rep_bern()
ps <- prepared_simulations()
rep_sims <- repeatable(run_simulations, seed = 122) # setting seed to ensure repeatable outcomes
x <- rep_sims(ps, bs)
return(x)
})
}
