% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability_functions.R
\name{sample_excess_kurtosis}
\alias{sample_excess_kurtosis}
\title{Sample excess kurtosis}
\usage{
sample_excess_kurtosis(x)
}
\arguments{
\item{x}{Sample data.}
}
\value{
A number.
}
\description{
This function calculates the samples excess kurtosis, adjusted for sample size.
}
\examples{
sample_exkurtosis <- sample_excess_kurtosis(data)
}
