% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{sim_bern}
\alias{sim_bern}
\title{sim_bern}
\usage{
sim_bern(the_right_data = NULL)
}
\arguments{
\item{the_right_data}{Data frame, which is either the scaled, detrended or core data, depending on the inputs selected}
}
\value{
Data frame with 15,000 simulated values for each peril.
}
\description{
This function simulates the frequency data based on the mle of bernoulli.
}
\examples{
simulate_bernoulli <- reactive({
rd <- get_right_data()
rd <- rd[[2]]
temp<- sim_bern(rd)
return(temp)
})
}
