% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_manipulation.R
\name{sum_across_two_indexes}
\alias{sum_across_two_indexes}
\title{Sum across two indexes}
\usage{
sum_across_two_indexes(
  list_structure,
  first_level_indexes,
  second_level_indexes
)
}
\arguments{
\item{second_level_indexes}{}
}
\description{
This function subsets a list of lists to sum across in the format sum(x[[i_1]][[j_1]] + x[[i_2]][[j_2]]). Used within the combine_freq_sev_ci function.
}
\examples{
sum_across_two_indexes(list_structure = freq_sev_ci, first_level_indexes = peril_names,second_level_indexes = peril_distributions)
}
