% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tool_formatting.R
\name{tab_maker}
\alias{tab_maker}
\title{tab_maker}
\usage{
tab_maker(n = 1, label = "DATA SELECTION", input, tab_data)
}
\arguments{
\item{n}{Tab number.}

\item{label}{Tab name.}

\item{input}{Tab inputs.}

\item{tab_data}{Tab data.}
}
\value{
UIs for panel headers.
}
\description{
Function for conditional colouring, etc. in the tab panel headers.
}
\examples{
output$tool_settings_ui <- renderUI({
tab_maker(n = 1, label = 'DATA SELECTION',
          input = input,
          tab_data = tab_data)
})
output$data_ui <- renderUI({
  tab_maker(n = 2, label = 'DATA MANIPULATION',
            input = input,
            tab_data = tab_data)
})
output$input_ui <- renderUI({
  tab_maker(n = 3, label = 'SIMULATIONS',
            input = input,
            tab_data = tab_data)
})
}
