% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_manipulation.R
\name{transform_core_data}
\alias{transform_core_data}
\title{Transform Core data}
\usage{
transform_core_data(cdx)
}
\arguments{
\item{cdx}{}
}
\value{
Data frame
}
\description{
Function for making core data a one row per peril-year data set
}
\examples{
out <- dat
out <- transform_core_data(out)
}
