% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_manipulation.R
\name{values_to_frequency}
\alias{values_to_frequency}
\title{Values to frequency}
\usage{
values_to_frequency(cdx)
}
\arguments{
\item{cdx}{}
}
\value{
Data frame
}
\description{
This function converts values to frequency data.
}
\examples{
cdx1 <- elongate_core_data(cdx1)
cdx2 <- values_to_frequency(cdx1)
}
