% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_manipulation.R
\name{widen_core_data}
\alias{widen_core_data}
\title{Widen core data}
\usage{
widen_core_data(cdx)
}
\arguments{
\item{cdx}{}
}
\value{
Data frame
}
\description{
Function widens the data from long to wide format.
}
\examples{
if(nrow(data) <=3){
the_data  <- tibble(` ` = "Not enough observations available")
} else {
  the_data <- widen_core_data(data)
  }
}
